requirements.txt
fastapi
python-docx
anthropic
pytesseract
pymupdf
python-dotenv
python-multipart
uvicorn
pandas
openpyxl
openai
boto3